refactor: move table redefined data formatter code
